refactor: remove redundant hooks
